Separate queries and mutations in their own file
Each query and mutation have their own file, under the `api` folder.
